combine spoon and paper module
