+++ paper/build.gradle.kts
@@ -1,5 +1,4 @@
 plugins {
-    alias(libs.plugins.paperweight)
     alias(libs.plugins.runPaper)
     alias(libs.plugins.shadow)
 
@@ -19,15 +18,15 @@ repositories {
 }
 
 dependencies {
-    paperweight.paperDevBundle(libs.versions.paper)
-
     implementation(projects.crazycratesCore)
 
     implementation(libs.triumph.cmds)
 
-    implementation(libs.fusion.paper) {
-        exclude("org.yaml")
-    }
+    implementation(libs.fusion.paper)
+
+    compileOnly(libs.hikari.cp)
+
+    compileOnly(libs.paper)
 
     implementation(libs.metrics)
 
@@ -37,6 +36,41 @@ dependencies {
 }
 
 tasks {
+    shadowJar {
+        archiveBaseName.set(rootProject.name)
+        archiveClassifier.set("")
+
+        listOf(
+            "com.ryderbelserion"
+        ).forEach {
+            relocate(it, "libs.$it")
+        }
+    }
+
+    assemble {
+        dependsOn(shadowJar)
+
+        doLast {
+            copy {
+                from(shadowJar.get())
+                into(rootProject.projectDir.resolve("jars"))
+            }
+        }
+    }
+
+    processResources {
+        inputs.properties("name" to rootProject.name)
+        inputs.properties("version" to project.version)
+        inputs.properties("group" to project.group)
+        inputs.properties("apiVersion" to libs.versions.minecraft.get())
+        inputs.properties("description" to project.description)
+        inputs.properties("website" to "https://modrinth.com/plugin/crazycrates")
+
+        filesMatching("paper-plugin.yml") {
+            expand(inputs.properties)
+        }
+    }
+
     runServer {
         jvmArgs("-Dnet.kyori.ansi.colorLevel=truecolor")
