+++ paper/src/main/java/com/badbones69/crazycrates/CrazyCrates.java
@@ -1,13 +1,15 @@
 package com.badbones69.crazycrates;
 
+import com.badbones69.crazycrates.api.enums.other.Plugins;
 import com.badbones69.crazycrates.common.Server;
 import com.badbones69.crazycrates.common.config.ConfigManager;
 import com.badbones69.crazycrates.common.config.impl.ConfigKeys;
+import com.badbones69.crazycrates.listeners.DataListener;
 import com.badbones69.crazycrates.listeners.crates.CrateInteractListener;
 import com.badbones69.crazycrates.listeners.items.PaperInteractListener;
+import com.badbones69.crazycrates.managers.data.DataManager;
 import com.badbones69.crazycrates.support.MetricsWrapper;
 import com.badbones69.crazycrates.utils.MiscUtils;
-import com.badbones69.crazycrates.commands.CommandManager;
 import com.badbones69.crazycrates.listeners.BrokeLocationsListener;
 import com.badbones69.crazycrates.listeners.CrateControlListener;
 import com.badbones69.crazycrates.listeners.MiscListener;
@@ -22,10 +24,10 @@ import com.badbones69.crazycrates.support.placeholders.PlaceholderAPISupport;
 import com.badbones69.crazycrates.managers.BukkitUserManager;
 import com.badbones69.crazycrates.managers.InventoryManager;
 import com.badbones69.crazycrates.tasks.crates.CrateManager;
-import com.ryderbelserion.vital.files.enums.FileType;
-import com.ryderbelserion.vital.paper.VitalPaper;
-import com.badbones69.crazycrates.api.enums.other.Plugins;
-import com.ryderbelserion.vital.utils.Methods;
+import com.ryderbelserion.FusionApi;
+import com.ryderbelserion.api.enums.FileType;
+import com.ryderbelserion.paper.files.FileManager;
+import com.ryderbelserion.util.Methods;
 import me.arcaniax.hdb.api.HeadDatabaseAPI;
 import org.bukkit.plugin.PluginManager;
 import org.bukkit.plugin.java.JavaPlugin;
@@ -36,30 +38,31 @@ import java.util.Locale;
 import java.util.Timer;
 import static com.badbones69.crazycrates.utils.MiscUtils.registerPermissions;
 
-@ApiStatus.Internal
 public class CrazyCrates extends JavaPlugin {
 
-    @ApiStatus.Internal
     public static CrazyCrates getPlugin() {
         return JavaPlugin.getPlugin(CrazyCrates.class);
     }
 
-    private final VitalPaper vital;
-    private final Timer timer;
     private final long startTime;
+    private final Timer timer;
 
     public CrazyCrates() {
-        this.startTime = System.nanoTime();
-
-        this.vital = new VitalPaper(this);
+        this.startTime = System.currentTimeMillis();
 
         this.timer = new Timer();
     }
 
-    private InventoryManager inventoryManager;
+    private final FusionApi api = FusionApi.get();
+
+    private FileManager fileManager;
+    private DataManager dataManager;
+
     private BukkitUserManager userManager;
+
+    private InventoryManager inventoryManager;
     private CrateManager crateManager;
-    private HeadDatabaseAPI api;
+    private HeadDatabaseAPI hdb;
 
     private Server instance;
 
@@ -70,7 +73,11 @@ public class CrazyCrates extends JavaPlugin {
         this.instance = new Server(getDataFolder());
         this.instance.apply();
 
-        this.vital.getFileManager().addFile("locations.yml", FileType.YAML).addFile("data.yml", FileType.YAML).addFile("respin-gui.yml", "guis", false, FileType.YAML)
+        this.api.enable(this);
+
+        this.fileManager = this.api.getFileManager();
+
+        this.fileManager.addFile("data.yml", FileType.YAML).addFile("locations.yml", FileType.YAML).addFile("respin-gui.yml", "guis", false, FileType.YAML)
                 .addFile("crates.log", "logs", false, FileType.NONE)
                 .addFile("keys.log", "logs", false, FileType.NONE)
                 .addFolder("crates", FileType.YAML)
@@ -83,11 +90,14 @@ public class CrazyCrates extends JavaPlugin {
         registerPermissions();
 
         if (Plugins.head_database.isEnabled()) {
-            this.api = new HeadDatabaseAPI();
+            this.hdb = new HeadDatabaseAPI();
         }
+
+        this.dataManager = new DataManager().init();
 
         this.inventoryManager = new InventoryManager();
         this.crateManager = new CrateManager();
+
         this.userManager = new BukkitUserManager();
 
         this.instance.setUserManager(this.userManager);
@@ -106,10 +116,9 @@ public class CrazyCrates extends JavaPlugin {
             this.metrics.start();
         }
 
-        // Load commands.
-        CommandManager.load();
+        final PluginManager pluginManager = this.getServer().getPluginManager();
 
-        final PluginManager manager = getServer().getPluginManager();
+        pluginManager.registerEvents(new DataListener(), this);
 
         List.of(
                 // Other listeners.
@@ -126,7 +135,7 @@ public class CrazyCrates extends JavaPlugin {
                 new CrateOpenListener(),
 
                 new PaperInteractListener()
-        ).forEach(listener -> manager.registerEvents(listener, this));
+        ).forEach(listener -> pluginManager.registerEvents(listener, this));
 
         this.crateManager.loadCustomItems();
 
@@ -152,6 +161,8 @@ public class CrazyCrates extends JavaPlugin {
 
     @Override
     public void onDisable() {
+        this.api.disable();
+
         // Cancel the tasks
         getServer().getGlobalRegionScheduler().cancelTasks(this);
         getServer().getAsyncScheduler().cancelTasks(this);
@@ -177,13 +188,11 @@ public class CrazyCrates extends JavaPlugin {
         MiscUtils.janitor();
     }
 
-    @ApiStatus.Internal
     public final InventoryManager getInventoryManager() {
         return this.inventoryManager;
     }
 
-    @ApiStatus.Internal
-    public final BukkitUserManager getUserManager() {
+    public BukkitUserManager getUserManager() {
         return this.userManager;
     }
 
@@ -193,12 +202,12 @@ public class CrazyCrates extends JavaPlugin {
     }
 
     @ApiStatus.Internal
-    public @Nullable final HeadDatabaseAPI getApi() {
-        if (this.api == null) {
+    public @Nullable final HeadDatabaseAPI getHdb() {
+        if (this.hdb == null) {
             return null;
         }
 
-        return this.api;
+        return this.hdb;
     }
 
     @ApiStatus.Internal
@@ -207,17 +216,25 @@ public class CrazyCrates extends JavaPlugin {
     }
 
     @ApiStatus.Internal
-    public @Nullable final MetricsWrapper getMetrics() {
+    public @Nullable
+    final MetricsWrapper getMetrics() {
         return this.metrics;
-    }
-
-    @ApiStatus.Internal
-    public final VitalPaper getVital() {
-        return this.vital;
     }
 
     @ApiStatus.Internal
     public final Timer getTimer() {
         return this.timer;
+    }
+
+    public FileManager getFileManager() {
+        return this.fileManager;
+    }
+
+    public DataManager getDataManager() {
+        return this.dataManager;
+    }
+
+    public FusionApi getApi() {
+        return this.api;
     }
 }
