+++ paper/src/main/java/com/badbones69/crazycrates/CrazyLoader.java
