+++ paper/src/main/java/com/badbones69/crazycrates/api/exception/CratesException.java
