+++ paper/src/main/java/com/badbones69/crazycrates/commands/crates/types/admin/crates/migrator/ICrateMigrator.java
@@ -3,7 +3,6 @@ package com.badbones69.crazycrates.commands.crates.types.admin.crates.migrator;
 import ch.jalu.configme.SettingsManager;
 import com.badbones69.crazycrates.CrazyCrates;
 import com.badbones69.crazycrates.api.enums.Messages;
-import com.badbones69.crazycrates.utils.ItemUtils;
 import com.badbones69.crazycrates.commands.crates.types.admin.crates.migrator.enums.MigrationType;
 import com.badbones69.crazycrates.common.config.ConfigManager;
 import com.badbones69.crazycrates.tasks.crates.CrateManager;
