+++ paper/src/main/java/com/badbones69/crazycrates/managers/InventoryManager.java
@@ -7,7 +7,7 @@ import com.badbones69.crazycrates.api.enums.other.keys.ItemKeys;
 import com.badbones69.crazycrates.api.objects.Crate;
 import com.badbones69.crazycrates.api.objects.Tier;
 import com.badbones69.crazycrates.api.builders.ItemBuilder;
-import com.ryderbelserion.vital.paper.api.builders.gui.types.PaginatedGui;
+import com.ryderbelserion.paper.builder.gui.types.PaginatedGui;
 import org.bukkit.entity.Player;
 import org.bukkit.event.inventory.InventoryCloseEvent;
 import org.bukkit.inventory.ItemStack;
