+++ paper/src/main/java/com/badbones69/crazycrates/managers/data/DataManager.java
@@ -1,8 +1,8 @@
-package com.badbones69.crazycrates.api.exception.data;
+package com.badbones69.crazycrates.managers.data;
 
 import com.badbones69.crazycrates.CrazyCrates;
-import com.badbones69.crazycrates.api.exception.data.interfaces.Connector;
-import com.badbones69.crazycrates.api.exception.data.types.SqliteConnector;
+import com.badbones69.crazycrates.managers.data.interfaces.Connector;
+import com.badbones69.crazycrates.managers.data.types.SqliteConnector;
 import java.io.File;
 
 public class DataManager {
