+++ paper/src/main/java/com/badbones69/crazycrates/tasks/menus/CrateAdminMenu.java
@@ -6,9 +6,9 @@ import com.badbones69.crazycrates.api.enums.Messages;
 import com.badbones69.crazycrates.api.enums.Permissions;
 import com.badbones69.crazycrates.utils.MiscUtils;
 import com.badbones69.crazycrates.common.config.impl.ConfigKeys;
-import com.ryderbelserion.vital.paper.api.builders.gui.interfaces.GuiFiller;
-import com.ryderbelserion.vital.paper.api.builders.gui.interfaces.GuiItem;
-import com.ryderbelserion.vital.paper.api.builders.gui.types.PaginatedGui;
+import com.ryderbelserion.paper.builder.gui.interfaces.GuiFiller;
+import com.ryderbelserion.paper.builder.gui.interfaces.GuiItem;
+import com.ryderbelserion.paper.builder.gui.types.PaginatedGui;
 import org.bukkit.Sound;
 import org.bukkit.entity.Player;
 import org.bukkit.event.inventory.ClickType;
