+++ paper/src/main/java/com/badbones69/crazycrates/tasks/menus/CratePreviewMenu.java
@@ -3,11 +3,10 @@ package com.badbones69.crazycrates.tasks.menus;
 import com.badbones69.crazycrates.api.builders.gui.DynamicInventoryBuilder;
 import com.badbones69.crazycrates.api.objects.Crate;
 import com.badbones69.crazycrates.api.objects.Tier;
-import com.ryderbelserion.vital.paper.api.builders.gui.interfaces.GuiFiller;
-import com.ryderbelserion.vital.paper.api.builders.gui.interfaces.GuiItem;
-import com.ryderbelserion.vital.paper.api.builders.gui.types.PaginatedGui;
+import com.ryderbelserion.paper.builder.gui.interfaces.GuiFiller;
+import com.ryderbelserion.paper.builder.gui.interfaces.GuiItem;
+import com.ryderbelserion.paper.builder.gui.types.PaginatedGui;
 import org.bukkit.entity.Player;
-
 import java.util.UUID;
 
 public class CratePreviewMenu extends DynamicInventoryBuilder {
