+++ paper/src/main/java/com/badbones69/crazycrates/tasks/menus/CrateTierMenu.java
@@ -4,9 +4,9 @@ import com.badbones69.crazycrates.api.builders.gui.StaticInventoryBuilder;
 import com.badbones69.crazycrates.api.enums.other.keys.ItemKeys;
 import com.badbones69.crazycrates.api.objects.Crate;
 import com.badbones69.crazycrates.api.objects.Tier;
-import com.ryderbelserion.vital.paper.api.builders.gui.interfaces.Gui;
-import com.ryderbelserion.vital.paper.api.builders.gui.interfaces.GuiFiller;
-import com.ryderbelserion.vital.paper.api.builders.gui.interfaces.GuiItem;
+import com.ryderbelserion.paper.builder.gui.interfaces.Gui;
+import com.ryderbelserion.paper.builder.gui.interfaces.GuiFiller;
+import com.ryderbelserion.paper.builder.gui.interfaces.GuiItem;
 import io.papermc.paper.persistence.PersistentDataContainerView;
 import net.kyori.adventure.sound.Sound;
 import org.bukkit.entity.Player;
