+++ paper/src/main/java/com/badbones69/crazycrates/utils/MsgUtils.java
@@ -1,9 +1,10 @@
 package com.badbones69.crazycrates.utils;
 
+import com.badbones69.crazycrates.common.config.ConfigManager;
+import com.badbones69.crazycrates.common.config.impl.ConfigKeys;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 import org.jetbrains.annotations.NotNull;
-import static com.badbones69.crazycrates.common.utils.CrazyUtil.getPrefix;
 import static java.util.regex.Matcher.quoteReplacement;
 
 public class MsgUtils {
@@ -11,7 +12,7 @@ public class MsgUtils {
     public static void sendMessage(CommandSender commandSender, @NotNull final String message, final boolean prefixToggle) {
         if (message.isEmpty()) return;
 
-        final String prefix = getPrefix();
+        final String prefix = ConfigManager.getConfig().getProperty(ConfigKeys.command_prefix);
 
         final boolean sendPrefix = !prefix.isEmpty() && prefixToggle;
