+++ paper/src/main/resources/paper-plugin.yml
@@ -6,14 +6,16 @@ api-version: '${apiVersion}'
 description: '${description}'
 website: '${website}'
 
+loader: '${group}.CrazyLoader'
+
 folia-supported: true
 
 dependencies:
   server:
     DecentHolograms:
-     load: BEFORE
-     required: false
-     join-classpath: true
+      load: BEFORE
+      required: false
+      join-classpath: true
     FancyHolograms:
       load: BEFORE
       required: false
@@ -23,17 +25,17 @@ dependencies:
       required: false
       join-classpath: true
     PlaceholderAPI:
-     load: BEFORE
-     required: false
-     join-classpath: true
+      load: BEFORE
+      required: false
+      join-classpath: true
     ItemsAdder:
       load: BEFORE
       required: false
       join-classpath: true
     Oraxen:
-     load: BEFORE
-     required: false
-     join-classpath: true
+      load: BEFORE
+      required: false
+      join-classpath: true
     Nexo:
       load: BEFORE
       required: false
