--- spoon/src/main/resources/paper-plugin.yml
@@ -1,52 +1,0 @@
-name: '${name}'
-main: '${group}.CrazyCrates'
-
-version: '${version}'
-api-version: '${apiVersion}'
-description: '${description}'
-website: '${website}'
-
-loader: '${group}.CrazyLoader'
-
-folia-supported: true
-
-dependencies:
-  server:
-    DecentHolograms:
-     load: BEFORE
-     required: false
-     join-classpath: true
-    FancyHolograms:
-      load: BEFORE
-      required: false
-      join-classpath: true
-    CMI:
-      load: BEFORE
-      required: false
-      join-classpath: true
-    PlaceholderAPI:
-     load: BEFORE
-     required: false
-     join-classpath: true
-    ItemsAdder:
-      load: BEFORE
-      required: false
-      join-classpath: true
-    Oraxen:
-     load: BEFORE
-     required: false
-     join-classpath: true
-    Nexo:
-      load: BEFORE
-      required: false
-      join-classpath: true
-    HeadDatabase:
-      load: BEFORE
-      required: false
-      join-classpath: true
-
-    # Other crate plugins to migrate for people
-    ExcellentCrates:
-      load: BEFORE
-      required: false
-      join-classpath: true
